requirements.txt
ipcalc==1.99.0
PyYAML==6.0